* commit '33c6cdaa175f20b1b326bc135ce6ebdedd899b24': (676 commits)
  EMS - Translator - Bugfix (reverting buggy code improvement)
  MetaSolver - Changes in order to address Pawel's comment
  EMS+MetaSolver - Chnages in order to address Pawel's comments for pull request
  EMS+MetaSolver - Removed not needed files
  added code to delete datasource
  EMS - Broker-CEP - Bugfix of expired event removal in CepEvalAggregator.
  EMS - Added workaround for MEL-724. It adds a metric for each top-level metric context.
  .
  MetaSolver - Modified a log message in CpModelHelper (minor)
  MetaSolver - Minor change in copyVarValuesFromDeployedSolution() logging level
  MetaSolver - Change in order to address MEL-728
  MetaSolver - Bugfix: added metric variables update in cp model, when reconfiguration is to be requested
  MetaSolver - Bugfix: added metric variables update in cp model, in /updateSolution endpoint
  MEL-645	Merge of S2D and Adapter - p6
  EMS - Translator - Added prettifyG2R() in order to print G2R rules map in a prettier format
  move logging dlms utility to debug
  MEL-729 - NPE in the Adapter during deleting monitors
  log level changed
  EMS+MetaSolver - Changes wrt MEL-724 (copy Var.Value from solution to corresponding Metric in CP model. E.g. AppActCardinality)
  MEL-645	Merge of S2D and Adapter - p5
  ...

+++ melodic-cache/src/main/java/eu/melodic/cache/NodeCandidatePredicates.java
@@ -42,12 +42,12 @@ public final class NodeCandidatePredicates {
         };
     }
 
-    public static Predicate<NodeCandidate> getStoragePredicate(Double value) {
+    public static Predicate<NodeCandidate> getStoragePredicate(Integer value) {
         Objects.requireNonNull(value);
         return new Predicate<NodeCandidate>() {
             @Override
             public boolean test(NodeCandidate nodeCandidate) {
-                return value.equals(nodeCandidate.getHardware().getDisk());
+                return value.equals(nodeCandidate.getHardware().getDisk().intValue());
             }
 
             @Override
